requirements.txt
absl-py==0.8.0
astor==0.8.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.1
opt-einsum==3.0.1
pkg-resources==0.0.0
protobuf==3.9.1
six==1.12.0
tb-nightly==1.15.0a20190806
tensorflow-gpu==2.0.0rc0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019080601
Werkzeug==0.15.6
wrapt==1.11.2
